prediction of sales of toys over 5.7M rupees
One toy store needs to generate a 5.7М rupees profit. Another way we need to close it. so here we used basic python tools for filtering, counting and finding mean also we had made a prediction: how much profit we can generate if we will have those categories in stock. and calculated the total additional profit for all the categories.